phase1/web/requirements.txt
Flask==3.0.0
pymongo==4.6.1
neo4j==5.15.0
Werkzeug==3.0.1
kafka-python==2.0.2
requests==2.31.0
opensearch-py==2.4.2
joblib==1.3.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4